increase timeout for k8s object update, add diagnostic output
the 5-second timeout for k8s reaching the target generation number was
seen to expire. This may be either because it is insufficient when the
control subsystem of k8s is under load or due to intervening
modification from outside of the adjust driver.

This change increases the timeout and adds a log message to detect
possible 3rd-party changes.