handle concurrent modifications to the target deployment(s)
- while waiting for update, stop the wait if generation>=expected
(before this change, we waited for generation==expected)
- if generation>expected, we know there was something that made a change
shortly after our change; in this case, the retrieved deployment data is
compared to the patch applied and if there is a mismatch, adjust fails
with 'reason=overwritten', otherwise success is returned.

When checking a deployment that was modified concurrently, only the
following are verified:
- resources/requests/[memory|cpu]
- resources/limits/[memory|cpu]
(and, only those that 'adjust' actually attempted to modify are checked)

Therefore, any other changes to the deployment (including 'replicas'!)
are OK, we will complain only if our mem/cpu adjustment was overwritten.